+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add ability to extract trace headers (traceparent and tracestate) from Xmidt headers, JSON, and msgpack and propagates the trace context accordingly [#115](https://github.com/xmidt-org/candlelight/pull/115)
 
 ## [v0.0.15]
 - Split up the tracing configuration options so user can decide exactly what gets traced and when [#104](https://github.com/xmidt-org/candlelight/pull/104)